\name{invlogit}
\alias{invlogit}
\title{invlogit}
\description{Transforming (-Inf,Inf) to original scale (0,1)}
\usage{invlogit(q, ...)}
\arguments{
  \item{q}{
}
  \item{\dots}{
}
}
\details{function \eqn{f(z) = \frac{e^{z}}{e^{z} + 1} \! = \frac{1}{1 + e^{-z}} \!}}


\author{Thomas Wutzler}



\seealso{\code{\link{logit}}

\code{\link{logitnorm}}}

