\name{logitnorm-package}
\alias{logitnorm-package}
\docType{package}
\title{Functions for the logitnormal distribution.}
\description{Density, distribution, quantile and random generation function for the logitnormal distribution. Estimation of the mode and the first two moments. Estimation of distribution parameters.}
\details{
\tabular{ll}{Package: \tab logitnorm\cr
Title: \tab Functions for the logitnormal distribution.\cr
Version: \tab 0.8.22\cr
Date: \tab 2012-03-07\cr
Author: \tab Thomas Wutzler\cr
Maintainer: \tab Thomas Wutzler <twutz@bgc-jena.mpg.de>\cr
Depends: \tab \cr
Suggests: \tab inlinedocs,debug\cr
License: \tab GPL-2\cr
LazyLoad: \tab yes\cr}
}
\author{Thomas Wutzler}

\keyword{ package }


