\name{plogitnorm}
\alias{plogitnorm}
\title{plogitnorm}
\description{Distribution function for logitnormal distribution	}
\usage{plogitnorm(q, mu = 0, sigma = 1, ...)}
\arguments{
  \item{q}{vector of quantiles}
  \item{mu}{location distribution parameter}
  \item{sigma}{scale distribution parameter}
  \item{\dots}{further arguments to pnorm}
}



\author{Thomas Wutzler}



\seealso{\code{\link{logitnorm}}}

