\name{qlogitnorm}
\alias{qlogitnorm}
\title{qlogitnorm}
\description{Quantiles of logitnormal distribution.	}
\usage{qlogitnorm(p, mu = 0, sigma = 1, ...)}
\arguments{
  \item{p}{vector of probabilities}
  \item{mu}{location distribution parameter}
  \item{sigma}{scale distribution parameter}
  \item{\dots}{further arguments to plogis}
}



\author{Thomas Wutzler}



\seealso{\code{\link{logitnorm}}}

