\name{rlogitnorm}
\alias{rlogitnorm}
\title{rlogitnorm}
\description{Random number generation for logitnormal distribution}
\usage{rlogitnorm(n, mu = 0, sigma = 1, ...)}
\arguments{
  \item{n}{number of observations}
  \item{mu}{distribution parameter}
  \item{sigma}{distribution parameter}
  \item{\dots}{arguments to \code{\link{rnorm}}}
}



\author{Thomas Wutzler}



\seealso{\code{\link{logitnorm}}}

