\name{twCoefLogitnormMLEFlat}
\alias{twCoefLogitnormMLEFlat}
\title{twCoefLogitnormMLEFlat}
\description{Estimating coefficients of a maximally flat unimodal logitnormal distribution given the mode 	}
\usage{twCoefLogitnormMLEFlat(mle)}
\arguments{
  \item{mle}{numeric vector: the mode of the density function}
}



\author{Thomas Wutzler}





