\name{logitnorm-package}
\alias{logitnorm-package}
\title{Functions for the logitnormal distribution.}
\description{Density, distribution, quantile and random generation function for the logitnormal distribution. Estimation of the mode and the first two moments. Estimation of distribution parameters.}

\author{Thomas Wutzler <twutz@bgc-jena.mpg.de>}

\keyword{ package }


