\name{modeLogitnorm}
\alias{modeLogitnorm}
\title{modeLogitnorm}
\description{Mode of the logitnormal distribution by numerical optimization}
\usage{
modeLogitnorm(mu, sigma, tol = invlogit(mu)/1000)
}
\arguments{
  \item{mu}{parameter mu}
  \item{sigma}{parameter sigma}
  \item{tol}{precisions of the estimate}
}



\author{Thomas Wutzler <twutz@bgc-jena.mpg.de>}



\seealso{\code{\link{logitnorm}}}

