\name{plogitnorm}
\alias{plogitnorm}
\title{plogitnorm}
\description{Distribution function for logitnormal distribution	}
\usage{
plogitnorm(q, mu = 0, sigma = 1, ...)
}
\arguments{
  \item{q}{
}
  \item{mu}{distribution parameters}
  \item{sigma}{
}
  \item{\dots}{
}
}



\author{Thomas Wutzler <twutz@bgc-jena.mpg.de>}



\seealso{\code{\link{logitnorm}}}

