\name{qlogitnorm}
\alias{qlogitnorm}
\title{qlogitnorm}
\description{Quantiles of logitnormal distribution.	}
\usage{
qlogitnorm(p, mu = 0, sigma = 1, ...)
}
\arguments{
  \item{p}{
}
  \item{mu}{distribution parameters}
  \item{sigma}{
}
  \item{\dots}{
}
}



\author{Thomas Wutzler <twutz@bgc-jena.mpg.de>}



\seealso{\code{\link{logitnorm}}}

