\name{rlogitnorm}
\alias{rlogitnorm}
\title{rlogitnorm}
\description{Random number generation for logitnormal distribution}
\usage{rlogitnorm(mu = 0, sigma = 1, ...)}
\arguments{
  \item{mu}{distribution parameters}
  \item{sigma}{
}
  \item{\dots}{arguments to \code{\link{rnorm}}}
}



\author{Thomas Wutzler}



\seealso{\code{\link{logitnorm}}}

